requirements.txt
pytesseract
pillow
pynput
pydirectinput
pyautogui
